% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUDC.R
\name{AUDC}
\alias{AUDC}
\title{Area under Decision Curve}
\usage{
AUDC(x)
}
\arguments{
\item{x}{results of dca() function}
}
\value{
Area under decision curves for each model.
}
\description{
Area under Decision Curve
}
