% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dca.cph.R
\name{ggplot.dca.cph}
\alias{ggplot.dca.cph}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{dca.cph}(data, color = TRUE, linetype = TRUE, lwd = 1.05)
}
\arguments{
\item{data}{results of dca() function}

\item{color}{logical, whether models will be classified by color}

\item{linetype}{logical, whether models will be classified by line type}

\item{lwd}{line width}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
