% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dca.lrm.R
\name{ggplot.dca.lrm}
\alias{ggplot.dca.lrm}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{dca.lrm}(data, color = TRUE, linetype = TRUE)
}
\arguments{
\item{data}{results of dca() function}

\item{color}{logical, whether models will be classified by color}

\item{linetype}{logical, whether models will be classified by line type}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
\examples{
1+1
}
