% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.dca.R
\name{ggplot.ggdca}
\alias{ggplot.ggdca}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{ggdca}(data)
}
\arguments{
\item{data}{results of dca() function}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
\examples{
library(rms)
library(ggDCA)

# one model
model1 <- lrm(status~age,lung)
dt=dca(model1)
ggplot(dt)

# 2 or more model
model1 <- lrm(status~age,lung)
model2 <- lrm(status~age+sex,lung)
dt=dca(model1,model2)
ggplot(dt)
}
