% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.rFP.p100.R, R/ggplot_dca.cph.R,
%   R/ggplot_dca.lrm.R
\name{ggplot.rFP.p100}
\alias{ggplot.rFP.p100}
\alias{ggplot.dca.cph}
\alias{ggplot}
\alias{ggplot.dca.lrm}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{rFP.p100}(
  data,
  mapping,
  color = TRUE,
  linetype = TRUE,
  lwd = 1.05,
  ...,
  environment = parent.frame()
)

\method{ggplot}{dca.cph}(
  data,
  mapping,
  color = TRUE,
  linetype = TRUE,
  lwd = 1.05,
  ...,
  environment = parent.frame()
)

\method{ggplot}{dca.lrm}(
  data,
  mapping,
  color = TRUE,
  linetype = TRUE,
  lwd = 1.05,
  ...,
  environment = parent.frame()
)
}
\arguments{
\item{data}{results of dca() function}

\item{mapping}{ignore}

\item{color}{logical, whether models will be classified by color}

\item{linetype}{logical, whether models will be classified by line type}

\item{lwd}{line width}

\item{...}{ignore}

\item{environment}{ignore}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
