% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.rFP.p100.R
\name{ggplot.rFP.p100}
\alias{ggplot.rFP.p100}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{rFP.p100}(data, color = TRUE, linetype = TRUE)
}
\arguments{
\item{data}{results of dca() function}

\item{color}{logical, whether models will be classified by color}

\item{linetype}{logical, whether models will be classified by line type}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
