% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIRI.R
\docType{data}
\name{LIRI}
\alias{LIRI}
\title{ICGC Liver Data from Japan}
\format{
An object of class \code{data.frame} with 232 rows and 6 columns.
}
\usage{
data(LIRI)
}
\description{
This data is a liver cancer data from Japan Data released in
ICGC database (\href{https://dcc.icgc.org/releases/release_24/Projects/LIRI-JP}{Link}).
It cantains time, event and four genes.
}
\examples{
data(LIRI)
}
\keyword{datasets}
