% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\alias{range.dca.lrm}
\alias{range.dca.cph}
\title{Ranges for net benefit}
\usage{
\method{range}{dca.lrm}(...)

\method{range}{dca.cph}(...)
}
\arguments{
\item{...}{results of dca() function}
}
\value{
a dataframe contains the minium and maximum of net benefit for each model
}
\description{
Ranges for net benefit
}
