% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFP.p100.R
\name{rFP.p100}
\alias{rFP.p100}
\title{Calculate reduction in false positive count}
\usage{
rFP.p100(x)
}
\arguments{
\item{x}{result of dca() function}
}
\value{
a dataframe contains thresholds, rFP.100: reduction
in false positive count per 100 patients.
}
\description{
Calculate reduction in false positive count
}
