% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca}
\alias{dca}
\title{Calculate Decision Curve Data}
\usage{
dca(..., model.names = do::get_names(...), include.cutoff = TRUE)
}
\arguments{
\item{...}{one or more results of logistic or cox regression}

\item{model.names}{names for models}

\item{include.cutoff}{logical. Whether the cutoff threshold should be included.}
}
\value{

}
\description{
Calculate Decision Curve Data
}
\examples{
1+1
}
