% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca}
\alias{dca}
\title{Calculate Decision Curve Data}
\usage{
dca(..., model.names = do::get_names(...), test.harm = 0, times = "median")
}
\arguments{
\item{...}{one or more results of logistic or cox regression}

\item{model.names}{names for models}

\item{test.harm}{test harm, default is 0}

\item{times}{times for cox regresion, default is 'median'}
}
\value{
a dataframe contains thresholds, TPR: true positive rate, FPR: false
positive rate, NB: net benefit, model: model names.
}
\description{
Calculate Decision Curve Data
}
\references{
Vickers, A. J., & Elkin, E. B. (2006). Decision Curve Analysis: A Novel Method for Evaluating Prediction Models. Medical Decision Making, 26(6), 565–574. https://doi.org/10.1177/0272989X06295361
}
