% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dca.lrm.R
\name{ggplot.dca.lrm}
\alias{ggplot.dca.lrm}
\title{Plot Decision Curve}
\usage{
\method{ggplot}{dca.lrm}(data, color = TRUE, linetype = TRUE, lwd = 1.05)
}
\arguments{
\item{data}{results of dca() function}

\item{color}{logical, whether models will be classified by color}

\item{linetype}{logical, whether models will be classified by line type}

\item{lwd}{line width}
}
\value{
a ggplot2 picture
}
\description{
Plot Decision Curve
}
\examples{
library(rmda)
library(ggDCA)
library(rms)
data(dcaData)

base.model <- lrm(Cancer~Age + Female + Smokes,data = dcaData)

d <-dca(base.model)

ggplot(d)

\donttest{
full.model <- lrm(Cancer~Age + Female + Smokes + Marker1 + Marker2,
                  data = dcaData)

d <- dca(base.model,full.model)

ggplot(d)

ggplot(d,color=FALSE)

ggplot(d,linetype = FALSE)

rfp=rFP.p100(x=d)

ggplot(rfp)
}
AUDC(d)
range(d)

}
